+++ src/classes/utils/Client.d.ts
@@ -432,7 +432,7 @@ export class Client {
         z: number;
     }, stage: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10): void;
     resetCamera(): void;
-    cooldown(item: itemName, tickAmount?: number): void;
+    cooldown(item: itemName, seconds?: number): void;
     demo(message: demoMessage): void;
     elderGuardian(): void;
     win(showCredits: boolean): void;